+++ src/idc/filter/objdet/__init__.py
@@ -3,3 +3,4 @@ from ._coerce_mask import CoerceMask
 from ._dimension_discarder import DimensionDiscarder
 from ._label_present import LabelPresent
 from ._polygon_discarder import PolygonDiscarder
+from ._polygon_simplifier import PolygonSimplifier