+++ CHANGES.rst
@@ -7,6 +7,7 @@ Changelog
 - added generic plugins that take user Python functions: `from-pyfunc`, `pyfunc-filter`, `to-pyfunc`
 - added `idc-exec` tool that uses generator to produce variable/value pairs that are used to expand
   the provided pipeline template which then gets executed
+- added `polygon-simplifier` filter for reducing number of points in polygons
 
 
 0.0.1 (2024-05-06)